+++ TP_Pokemon_Console/firepokemon.cpp
@@ -1,6 +1,6 @@
 #include "firepokemon.h"
 
-FirePokemon::FirePokemon(const std::string &itsName, float itsSize, float itsWeight, int itsHealthPoint, int itsStrengthPower, int itsPaws)
+FirePokemon::FirePokemon(const std::string &itsName, float itsSize, float itsWeight, float itsHealthPoint, int itsStrengthPower, int itsPaws)
     : Pokemon(itsName, itsSize, itsWeight, itsHealthPoint, itsStrengthPower),
       itsPaws(itsPaws)
 {}
